Version v0.3.0
Major change since version v0.2.1 is upgrade to Hyper 0.12.x
which also requires API changes to Restson.

Header type is no longer available in Hyper so the option to use
it in Restson API was also removed. set_header_raw was also
renamed to set_header because it is now the only interface to
set headers. Moreover, Result was added to set_header to
indicate if the header contains invalid characters.

Since the public API changes, the minor version is increased.
Other changes required for Hyper update are internal and should
not affect the library user.